requirements.txt
flask
pillow
pytesseract
requests
gunicorn
flask_sqlalchemy
flask_bcrypt
mysqlclient
pymysql
elevenlabs